Improve search dict performance (#67)
* Implement unit tests and benchmark tests for search_dict

The unit tests implemented with this commit aim to provide some
trust in the correctness of search_dict's implementation. The
benchmark test was done to provide some reference point for the
performance of other implementations to compare agains.

* Implement search_dict in terms of a stack instead of recursive calls

This commit aims to improve the performance of search_dict. To measure
the performance in at least some capacity a previous commit introduced
a test benchmark.

The idea behind the new implementation is to use a stack when searching
dictionaries instead of using recursive calls. The author would expect
this approach to be slightly faster since slapping something at the end
of a list is expected to be fast than allocating a new stack frame for
a function call. The benchmark suggests that the new implementation to
be 25% faster then the old one.